[KMP] refactor: make HiddenNavBarBottomSheetSample a private function in androidMain
Remove the expect-actual pattern for HiddenNavBarBottomSheetSample. The
function is now a standalone private composable in androidMain since
hideNavigationBar is an Android-only BottomSheet parameter. The shared
mobileMain display no longer includes the Hidden Navigation Bar section.
