+++ kmp/composeApp/src/mobileMain/kotlin/com/teya/lemonade/BottomSheetDisplay.kt
@@ -22,58 +22,6 @@ import com.teya.lemonade.core.LemonadeButtonSize
 import com.teya.lemonade.core.LemonadeButtonVariant
 
 @Composable
-private fun HiddenNavBarBottomSheetSample() {
-    var showSheet by remember { mutableStateOf(false) }
-
-    LemonadeUi.Text(
-        text = "Hides the system navigation bar inside the sheet dialog (Android only)",
-        textStyle = LemonadeTheme.typography.bodySmallRegular,
-        color = LemonadeTheme.colors.content.contentSecondary,
-    )
-    LemonadeUi.Button(
-        label = "Open Without Navigation Bar",
-        onClick = { showSheet = true },
-        variant = LemonadeButtonVariant.Secondary,
-        size = LemonadeButtonSize.Medium,
-    )
-
-    LemonadeUi.BottomSheet(
-        expanded = showSheet,
-        onDismissRequest = { showSheet = false },
-        hideNavigationBar = true,
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(
-                space = LemonadeTheme.spaces.spacing400,
-            ),
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(all = LemonadeTheme.spaces.spacing400),
-        ) {
-            LemonadeUi.Text(
-                text = "Hidden Navigation Bar",
-                textStyle = LemonadeTheme.typography.headingSmall,
-            )
-            LemonadeUi.Text(
-                text = "The system navigation bar is hidden in this bottom sheet. Swipe up from the bottom edge to temporarily reveal it.",
-                color = LemonadeTheme.colors.content.contentSecondary,
-            )
-            LemonadeUi.Button(
-                label = "Close",
-                onClick = { showSheet = false },
-                variant = LemonadeButtonVariant.Primary,
-                size = LemonadeButtonSize.Medium,
-            )
-            Spacer(
-                modifier = Modifier.height(
-                    height = LemonadeTheme.spaces.spacing400,
-                ),
-            )
-        }
-    }
-}
-
-@Composable
 internal fun BottomSheetSampleDisplay() {
     var showBasicSheet by remember { mutableStateOf(false) }
     var showNoDragHandleSheet by remember { mutableStateOf(false) }
@@ -95,11 +43,6 @@ internal fun BottomSheetSampleDisplay() {
                 variant = LemonadeButtonVariant.Secondary,
                 size = LemonadeButtonSize.Medium,
             )
-        }
-
-        // Hidden Navigation Bar (Android only)
-        BottomSheetSection(title = "Hidden Navigation Bar") {
-            HiddenNavBarBottomSheetSample()
         }
 
         // Without Drag Handle
